Fix how querystrings are dealt with for photon hosts, protcolless urls.
Querystrings for photon using hosts were previously jammed into the pathname.
This teaches the url generator to strip them and insert new params based on the
options passed.

It also adds support for protocolless urls, which are becoming pretty common.

Last, it makes the inception handling for previously photoned URLs more robust,
and adds a number of tests to validate that the inception detection works.